+++ runtime/lib/sign_extend_template_c.h
@@ -29,7 +29,7 @@ ASL_sign_extend_bits(M, N, int width, ASL_BITS_TYPE_M x, ASL_int_t n)
         int lead_limb = (width - 1) >> 6;
         int lead_msb = (width - 1) & 63;
         uint64_t lead = LIMB_M(x, lead_limb);
-        bool is_negative = (lead & (1 << lead_msb)) != 0;
+        bool is_negative = (lead & (((uint64_t)1) << lead_msb)) != 0;
 
         if (is_negative) {
                 int lead_res_limb = ((int)n + 63) >> 6;
